Improve the SloanDigitalSkyDataset class (#185)
This pull request overhauls the SloanDigitalSkyDataset class.
It adds the following:

    Allows loading multiple fields at once (#116) (#118) 
         (f1f037c, 454b40c)
    Allow for different stampsizes (1e0266e)
    Extract metadata about the star (50b04a1, 45391f1)
    Caching for the SDSS metadata FITS objects (54d309b)
    Load in the PSF from SDSS as well as the metadata 
        (6b0bb50,344c8a1, 0cf76d3)
    Add subpixel-centering method for the stamps (dfc6ec3)


* Added placeholder afnp code to case studies

* Extract fluxes and the pt in the SDSS class

* fixes issue 118 regarding loading multiple fields

If data for a field is missing, we show the error and set that
index equal to None

* Altering fix for missing field data

Instead of appending a None object, we now just don't append anything.

* make the loading message optional

* Added option to control the extracted stamp size

* now extract backgrounds and image widths

* Finished the caching mechanism in sdss class

This will now pickle the result of get_from_disk(), making reads much
faster. The "overwrite_cache" flag will re-load the data (necessary
if changes are made to the code)

* Don't add edge stars by def / append fluxes correctly

We no longer add edge stars by default. And if we exclude a
star its flux is no longer appended

* Added caching for the SDSS metadata objects as well

* Added subpixel correction to SDSS dataset

The sdss option now centers the stars by default using the grid_sample
function from PyTorch. This is a bilinear interpolation. This should
reduce the nuisance noise from the bias caused by subpixel differences
in the stars.

* Return the pr now in fetch_bright stars

* fixed the makefile for psField

* Added mechanism to load the PSF from the SDSS data

* Lazy-load SdssPSF Class to avoid reading file

This will avoiding reading the sdss PSF file unless the cache
of the objects is being recreated.

* Make the SDSS Psfs return as a numpy array

* Handle case where there are no bright stars

* Removed changed fits files

* Revert "Removed changed fits files"

This reverts commit 7e8a51ec4af1381ef2a663979609e36dc9c10e3d.

* removed placeholder afnp file

* Reformat the sdss.py file with black

* Added fitsio dependency to pyproject

* Update SDSS test to modified fields argument

* Change warning to info for fields which cannot be read

* allow for caches of sdss to be separate for stampsize

* Updated test for better coverage of SDSS:

* fixed black formatting issues

* Removed WIP psf caching code that isn't fully implemented

* Removed edge stamp code which was never used