Fix plotting bug in plot_locs()
This fixes plot_locs() to expect locations in "pixel" units.
Previously, plot_locs() was expecting normalized locations (between 0 and 1),
which it then multipled by the slen of the image.
With the change to the FullCatalog class, we were passing it
"ploc" (pixel locations), which are already scaled to the size
of the image.