prepare to train deblender (#8)
* refactor so that there is a single functiont that performs tile grid_sampling

I noticed that that the function that was shifting sources in padded tiles in `render_tiles` was exactly the same function as `center_tiles` in grid (that was being used by the deblender in binary encoders)! The only difference was one line, where the offsets were the negative of each other (makes sense, one was shift away from the center and one was shifting towards the center). Since this is the most complicated function in both of the modules, I decided to consolidate them both into a single function that uses grid_sample and adda  flag `center` which controls which mode to use. I also added unit tests that check that these functions (with the respective flag) are all equal to each other.

* just to be clear that the ptiles are flat in the
input here

* need not put this function at the beginning

* uneeded assertion

* simplify assertions one line

* removed the `_crop_loss_at_border` flag

As far as I can tell, this flag was always 'False' in the model in the `desc_mnras` branch and it was never changed in any of the config files. So I think it was never set to 'True' anywhere. Regardless, I never think we would like to ignore the loss in the border pading, as we want to evaluate reconstruction in full galaxies not just what's in the tile.

* validate border padding given

* assertionst to be very careful

* test on render_ptiles functions

* more realistic but shouldn't matter

* IMPORTANT CHANGE: loss function of deblender encoder

I realized that since we added galaxies in the padding, we need to add said padding to the reconstruction of the galaxy encoder when evaluating the loss function, otherwise the deblender will try to reconstruct the flux of galaxies that might be blended with galxies in the central square (as this will decrease the loss function). By adding the padding to the recon_mean it will penalize the encoder for over predicting flux and only target the galaxies in tiles, in the central square.

* trivial fix in one of the tests

* rename

* renames

* missed validation correct unpacking

* draft training script for deblender; added default values

* need to load state_dict...

* small bug

* correct path

* taking in flat tiles, bug

* BUG rearrange needs more context

* BUG rearrange missing context

* TST add more tests to check we are not actualy trimming!

* add stars to padding as well

* REF carefully remove all uses of extremly confusing fit_source_ptile function and its derivatives

I also added a new test in `test_render_ptiles` where I make it clear that that function was not doing anything in the case of a galaxy tiles. The real point where it was getting trimmed was in the grid_sample function, but that is confusing because the function that was using grid_sample was suppose to just shift sources but it ended up trimming padded tiles of galaxies too, but was it doing it correctly? I also removed the use in the measuing galaxy ellipticyt function because I can just create a PSF with the correct size.

* new tests

* REF renaming variable

* might not be very useful

* notebook to check finally if our grid_sample is doing the correct thing

* test confirms we can do this

* TST test to check that shifting function also trims in the case of 53x53

* REF ensure that the function `shift_sources_in_ptiles`can be used safely in the case of 53x53 images

we add a fairly large docstring to document this change in the function, but we mostly add some checks so we can safely use this function in the case that image is 52x52 or 53x53, again the notebook.

Finally, I add the `validate_border_padding` function here to avoid circular imports

* REF remove `validate_border_padding` function to avoid circular imports

* ensure correct usage of `shift_sources_in_ptiles` in binary encoder

* REF correct imports

* small error

* TST one more check of correctness

* REF prefer using `ptile_slen` to bp as input to functions; also we never trim the image in `reconstruct_image_from_ptiles` so remove (can do manually later if we want)

* REF propagate updates so now it works

* update tsts

* BUG fix where we need to pass a tensor in the case where we don't need the padding

* test we can run one loop of encoders

* new file to run deblender

* typos

* avoid overriding arguments (functional)